requirements.txt
# API Framework
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2
python-multipart>=0.0.5

# Document Processing
pdf2image>=1.16.0
pillow>=8.3.1
opencv-python>=4.5.3
pytesseract>=0.3.8
fpdf>=1.7.2
PyPDF2>=3.0.0
pikepdf>=5.0.0

# NLP and PII Detection
spacy>=3.2.0
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0

# Layout Analysis
numpy>=1.21.0
scikit-learn>=0.24.2

# OCR (optional enhancements)
# easyocr>=1.4.1

# Database
sqlalchemy>=1.4.23
alembic>=1.7.1
psycopg2-binary>=2.9.1

# Testing
pytest>=6.2.5
httpx>=0.18.2  # For TestClient
